@@ -11,16 +11,16 @@
     ],
     "minimum-stability": "dev",
     "require": {
-        "php": ">=7.1",
+        "php": ">=7.4",
         "ext-curl": "*",
         "ext-json": "*",
         "ext-mbstring": "*",
-        "guzzlehttp/guzzle": "6.3"
+        "guzzlehttp/guzzle": "7.3"
     },
     "require-dev": {
         "phpdocumentor/phpdocumentor": "2.*",
         "squizlabs/php_codesniffer": "3.*",
-        "wp-coding-standards/wpcs": "1.2.0",
+        "wp-coding-standards/wpcs": "1.2.*",
         "phing/phing": "2.*"
     },
     "scripts": {
@@ -31,7 +31,8 @@
     "autoload": {
         "psr-4": { 
             "Inc\\": "./inc",
-            "Irisnet\\APIV1\\Client\\" : "./ext/lib/" 
+            "Migrations\\": "./migrations",
+            "Irisnet\\APIV2\\Client\\" : "./ext/lib/" 
         }
     }
 }
