@@ -2,10 +2,10 @@
 /**
  * LicenseInfo
  *
- * PHP version 5
+ * PHP version 7.4
  *
  * @category Class
- * @package  Irisnet\APIV1\Client
+ * @package  Irisnet\APIV2\Client
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
  */
@@ -13,12 +13,12 @@
 /**
  * Irisnet API
  *
- * Artificial Intelligence (AI) for image- and video-processing in realtime. This is an interactive documentation meant to give a place were you can quickly look up the endpoints and their schemas, while also giving you the option to try things out yourself.  Listed below you'll see the available endpoints of the API that can be expanded by clicking on it. Each expanded endpoint lists the request parameter (if available) and the request body (if available). The request body can list some example bodies and the schema, explaining each model in detail. Additionally you'll find a 'Try it out' button where you can type in your custom parameters and custom body and execute that against the API. The responses section in the expanded endpoint lists the possible responses with their corresponding status codes. If you've executed an API call it will also show you the response from the server.  Underneath the endpoints you'll find the model schemas. These are the models used for the requests and responses. By clicking on the right arrow you can expand the model and it will show you a description of the model and the model parameters. For nested models you can keep clicking the right arrow to reveal further details on it.
+ * Artificial Intelligence (AI) for image- and video-processing in real-time. This is an interactive documentation where you can quickly look up the endpoints and their schemas, while having the opportunity to try things out for yourself.  In the list below, you can see the available endpoints of the API, which can be expanded by clicking on them. Each expanded endpoint lists the request parameters (if available) and the request body (if available). The request body can list some example bodies and the schema, explaining each model in detail.  Additionally you'll find a 'Try it out' button that allows you to enter your custom parameters and custom body and execute that against the API. <b>Be sure to enter your license key to authorize the requests before using this documentation interactively.</b>  The responses section in the expanded endpoint lists the possible responses with their corresponding status codes. If you've executed an API call it will also show you the response from the server.  Underneath the endpoints you'll find the model schemas. These are the models used for the requests and responses. If you click on the right arrow, you can expand the model and get a description of the model and the model parameters. For nested models, you can keep clicking the right arrow for further details.  Clicking the link below the title at the top of this page opens the [OpenAPI specification](https://swagger.io/specification/) (OAS3) in JSON format. The OAS3 Spec allows the generation of clients in many programming languages. There are several free client generators available that can be used to get started easily.
  *
- * The version of the OpenAPI document: v1
- * 
+ * The version of the OpenAPI document: v2
+ * Contact: info@irisnet.de
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 7.0.1
  */
 
 /**
@@ -27,23 +27,24 @@
  * Do not edit the class manually.
  */
 
-namespace Irisnet\APIV1\Client\Model;
+namespace Irisnet\APIV2\Client\Model;
 
 use \ArrayAccess;
-use \Irisnet\APIV1\Client\ObjectSerializer;
+use \Irisnet\APIV2\Client\ObjectSerializer;
 
 /**
  * LicenseInfo Class Doc Comment
  *
  * @category Class
- * @description Describes the current balance of the given license key. A key has a certain amount of credits that can be used for any kind of AI recognition. The license key is invalid, when all of the credits have been used.
- * @package  Irisnet\APIV1\Client
+ * @description Describes the current balance of the given license key. A key has a certain amount of credits that can be used for any kind of AI recognition. The license key is invalid, when all of the credits have been used, the license was disabled or expired.
+ * @package  Irisnet\APIV2\Client
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
+ * @implements \ArrayAccess<string, mixed>
  */
-class LicenseInfo implements ModelInterface, ArrayAccess
+class LicenseInfo implements ModelInterface, ArrayAccess, \JsonSerializable
 {
-    const DISCRIMINATOR = null;
+    public const DISCRIMINATOR = null;
 
     /**
       * The original name of the model.
@@ -60,19 +61,45 @@ class LicenseInfo implements ModelInterface, ArrayAccess
     protected static $openAPITypes = [
         'credits_used' => 'int',
         'credits_remaining' => 'int',
-        'total_credits' => 'int'
+        'total_credits' => 'int',
+        'license_key' => 'string',
+        'privileges' => 'array<string,string>'
     ];
 
     /**
       * Array of property to format mappings. Used for (de)serialization
       *
       * @var string[]
+      * @phpstan-var array<string, string|null>
+      * @psalm-var array<string, string|null>
       */
     protected static $openAPIFormats = [
         'credits_used' => 'int32',
         'credits_remaining' => 'int32',
-        'total_credits' => 'int32'
+        'total_credits' => 'int32',
+        'license_key' => null,
+        'privileges' => null
     ];
+
+    /**
+      * Array of nullable properties. Used for (de)serialization
+      *
+      * @var boolean[]
+      */
+    protected static array $openAPINullables = [
+        'credits_used' => false,
+		'credits_remaining' => false,
+		'total_credits' => false,
+		'license_key' => false,
+		'privileges' => false
+    ];
+
+    /**
+      * If a nullable field gets set to null, insert it here
+      *
+      * @var boolean[]
+      */
+    protected array $openAPINullablesSetToNull = [];
 
     /**
      * Array of property to type mappings. Used for (de)serialization
@@ -95,6 +122,58 @@ class LicenseInfo implements ModelInterface, ArrayAccess
     }
 
     /**
+     * Array of nullable properties
+     *
+     * @return array
+     */
+    protected static function openAPINullables(): array
+    {
+        return self::$openAPINullables;
+    }
+
+    /**
+     * Array of nullable field names deliberately set to null
+     *
+     * @return boolean[]
+     */
+    private function getOpenAPINullablesSetToNull(): array
+    {
+        return $this->openAPINullablesSetToNull;
+    }
+
+    /**
+     * Setter - Array of nullable field names deliberately set to null
+     *
+     * @param boolean[] $openAPINullablesSetToNull
+     */
+    private function setOpenAPINullablesSetToNull(array $openAPINullablesSetToNull): void
+    {
+        $this->openAPINullablesSetToNull = $openAPINullablesSetToNull;
+    }
+
+    /**
+     * Checks if a property is nullable
+     *
+     * @param string $property
+     * @return bool
+     */
+    public static function isNullable(string $property): bool
+    {
+        return self::openAPINullables()[$property] ?? false;
+    }
+
+    /**
+     * Checks if a nullable property is set to null.
+     *
+     * @param string $property
+     * @return bool
+     */
+    public function isNullableSetToNull(string $property): bool
+    {
+        return in_array($property, $this->getOpenAPINullablesSetToNull(), true);
+    }
+
+    /**
      * Array of attributes where the key is the local name,
      * and the value is the original name
      *
@@ -103,7 +182,9 @@ class LicenseInfo implements ModelInterface, ArrayAccess
     protected static $attributeMap = [
         'credits_used' => 'creditsUsed',
         'credits_remaining' => 'creditsRemaining',
-        'total_credits' => 'totalCredits'
+        'total_credits' => 'totalCredits',
+        'license_key' => 'licenseKey',
+        'privileges' => 'privileges'
     ];
 
     /**
@@ -114,7 +195,9 @@ class LicenseInfo implements ModelInterface, ArrayAccess
     protected static $setters = [
         'credits_used' => 'setCreditsUsed',
         'credits_remaining' => 'setCreditsRemaining',
-        'total_credits' => 'setTotalCredits'
+        'total_credits' => 'setTotalCredits',
+        'license_key' => 'setLicenseKey',
+        'privileges' => 'setPrivileges'
     ];
 
     /**
@@ -125,7 +208,9 @@ class LicenseInfo implements ModelInterface, ArrayAccess
     protected static $getters = [
         'credits_used' => 'getCreditsUsed',
         'credits_remaining' => 'getCreditsRemaining',
-        'total_credits' => 'getTotalCredits'
+        'total_credits' => 'getTotalCredits',
+        'license_key' => 'getLicenseKey',
+        'privileges' => 'getPrivileges'
     ];
 
     /**
@@ -169,9 +254,6 @@ class LicenseInfo implements ModelInterface, ArrayAccess
         return self::$openAPIModelName;
     }
 
-    
-
-    
 
     /**
      * Associative array for storing property values
@@ -188,9 +270,29 @@ class LicenseInfo implements ModelInterface, ArrayAccess
      */
     public function __construct(array $data = null)
     {
-        $this->container['credits_used'] = isset($data['credits_used']) ? $data['credits_used'] : null;
-        $this->container['credits_remaining'] = isset($data['credits_remaining']) ? $data['credits_remaining'] : null;
-        $this->container['total_credits'] = isset($data['total_credits']) ? $data['total_credits'] : null;
+        $this->setIfExists('credits_used', $data ?? [], null);
+        $this->setIfExists('credits_remaining', $data ?? [], null);
+        $this->setIfExists('total_credits', $data ?? [], null);
+        $this->setIfExists('license_key', $data ?? [], null);
+        $this->setIfExists('privileges', $data ?? [], null);
+    }
+
+    /**
+    * Sets $this->container[$variableName] to the given data or to the given default Value; if $variableName
+    * is nullable and its value is set to null in the $fields array, then mark it as "set to null" in the
+    * $this->openAPINullablesSetToNull array
+    *
+    * @param string $variableName
+    * @param array  $fields
+    * @param mixed  $defaultValue
+    */
+    private function setIfExists(string $variableName, array $fields, $defaultValue): void
+    {
+        if (self::isNullable($variableName) && array_key_exists($variableName, $fields) && is_null($fields[$variableName])) {
+            $this->openAPINullablesSetToNull[] = $variableName;
+        }
+
+        $this->container[$variableName] = $fields[$variableName] ?? $defaultValue;
     }
 
     /**
@@ -232,10 +334,13 @@ class LicenseInfo implements ModelInterface, ArrayAccess
      *
      * @param int|null $credits_used Credits used for the license key.
      *
-     * @return $this
+     * @return self
      */
     public function setCreditsUsed($credits_used)
     {
+        if (is_null($credits_used)) {
+            throw new \InvalidArgumentException('non-nullable credits_used cannot be null');
+        }
         $this->container['credits_used'] = $credits_used;
 
         return $this;
@@ -256,10 +361,13 @@ class LicenseInfo implements ModelInterface, ArrayAccess
      *
      * @param int|null $credits_remaining Credits remaining for the license key.
      *
-     * @return $this
+     * @return self
      */
     public function setCreditsRemaining($credits_remaining)
     {
+        if (is_null($credits_remaining)) {
+            throw new \InvalidArgumentException('non-nullable credits_remaining cannot be null');
+        }
         $this->container['credits_remaining'] = $credits_remaining;
 
         return $this;
@@ -278,13 +386,70 @@ class LicenseInfo implements ModelInterface, ArrayAccess
     /**
      * Sets total_credits
      *
-     * @param int|null $total_credits Total credits contained within the license.
+     * @param int|null $total_credits Total credits contained within the license key.
      *
-     * @return $this
+     * @return self
      */
     public function setTotalCredits($total_credits)
     {
+        if (is_null($total_credits)) {
+            throw new \InvalidArgumentException('non-nullable total_credits cannot be null');
+        }
         $this->container['total_credits'] = $total_credits;
+
+        return $this;
+    }
+
+    /**
+     * Gets license_key
+     *
+     * @return string|null
+     */
+    public function getLicenseKey()
+    {
+        return $this->container['license_key'];
+    }
+
+    /**
+     * Sets license_key
+     *
+     * @param string|null $license_key The license key
+     *
+     * @return self
+     */
+    public function setLicenseKey($license_key)
+    {
+        if (is_null($license_key)) {
+            throw new \InvalidArgumentException('non-nullable license_key cannot be null');
+        }
+        $this->container['license_key'] = $license_key;
+
+        return $this;
+    }
+
+    /**
+     * Gets privileges
+     *
+     * @return array<string,string>|null
+     */
+    public function getPrivileges()
+    {
+        return $this->container['privileges'];
+    }
+
+    /**
+     * Sets privileges
+     *
+     * @param array<string,string>|null $privileges A map of privileges
+     *
+     * @return self
+     */
+    public function setPrivileges($privileges)
+    {
+        if (is_null($privileges)) {
+            throw new \InvalidArgumentException('non-nullable privileges cannot be null');
+        }
+        $this->container['privileges'] = $privileges;
 
         return $this;
     }
@@ -295,7 +460,7 @@ class LicenseInfo implements ModelInterface, ArrayAccess
      *
      * @return boolean
      */
-    public function offsetExists($offset)
+    public function offsetExists($offset): bool
     {
         return isset($this->container[$offset]);
     }
@@ -305,22 +470,23 @@ class LicenseInfo implements ModelInterface, ArrayAccess
      *
      * @param integer $offset Offset
      *
-     * @return mixed
+     * @return mixed|null
      */
+    #[\ReturnTypeWillChange]
     public function offsetGet($offset)
     {
-        return isset($this->container[$offset]) ? $this->container[$offset] : null;
+        return $this->container[$offset] ?? null;
     }
 
     /**
      * Sets value based on offset.
      *
-     * @param integer $offset Offset
-     * @param mixed   $value  Value to be set
+     * @param int|null $offset Offset
+     * @param mixed    $value  Value to be set
      *
      * @return void
      */
-    public function offsetSet($offset, $value)
+    public function offsetSet($offset, $value): void
     {
         if (is_null($offset)) {
             $this->container[] = $value;
@@ -336,9 +502,22 @@ class LicenseInfo implements ModelInterface, ArrayAccess
      *
      * @return void
      */
-    public function offsetUnset($offset)
+    public function offsetUnset($offset): void
     {
         unset($this->container[$offset]);
+    }
+
+    /**
+     * Serializes the object to a value that can be serialized natively by json_encode().
+     * @link https://www.php.net/manual/en/jsonserializable.jsonserialize.php
+     *
+     * @return mixed Returns data which can be serialized by json_encode(), which is a value
+     * of any type other than a resource.
+     */
+    #[\ReturnTypeWillChange]
+    public function jsonSerialize()
+    {
+       return ObjectSerializer::sanitizeForSerialization($this);
     }
 
     /**
