@@ -2,10 +2,10 @@
 /**
  * ModelInterface
  *
- * PHP version 5
+ * PHP version 7.4
  *
  * @category Class
- * @package  Irisnet\APIV1\Client\Model
+ * @package  Irisnet\APIV2\Client\Model
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
  */
@@ -13,12 +13,12 @@
 /**
  * Irisnet API
  *
- * Artificial Intelligence (AI) for image- and video-processing in realtime. This is an interactive documentation meant to give a place were you can quickly look up the endpoints and their schemas, while also giving you the option to try things out yourself.  Listed below you'll see the available endpoints of the API that can be expanded by clicking on it. Each expanded endpoint lists the request parameter (if available) and the request body (if available). The request body can list some example bodies and the schema, explaining each model in detail. Additionally you'll find a 'Try it out' button where you can type in your custom parameters and custom body and execute that against the API. The responses section in the expanded endpoint lists the possible responses with their corresponding status codes. If you've executed an API call it will also show you the response from the server.  Underneath the endpoints you'll find the model schemas. These are the models used for the requests and responses. By clicking on the right arrow you can expand the model and it will show you a description of the model and the model parameters. For nested models you can keep clicking the right arrow to reveal further details on it.
+ * Artificial Intelligence (AI) for image- and video-processing in real-time. This is an interactive documentation where you can quickly look up the endpoints and their schemas, while having the opportunity to try things out for yourself.  In the list below, you can see the available endpoints of the API, which can be expanded by clicking on them. Each expanded endpoint lists the request parameters (if available) and the request body (if available). The request body can list some example bodies and the schema, explaining each model in detail.  Additionally you'll find a 'Try it out' button that allows you to enter your custom parameters and custom body and execute that against the API. <b>Be sure to enter your license key to authorize the requests before using this documentation interactively.</b>  The responses section in the expanded endpoint lists the possible responses with their corresponding status codes. If you've executed an API call it will also show you the response from the server.  Underneath the endpoints you'll find the model schemas. These are the models used for the requests and responses. If you click on the right arrow, you can expand the model and get a description of the model and the model parameters. For nested models, you can keep clicking the right arrow for further details.  Clicking the link below the title at the top of this page opens the [OpenAPI specification](https://swagger.io/specification/) (OAS3) in JSON format. The OAS3 Spec allows the generation of clients in many programming languages. There are several free client generators available that can be used to get started easily.
  *
- * The version of the OpenAPI document: v1
- * 
+ * The version of the OpenAPI document: v2
+ * Contact: info@irisnet.de
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 7.0.1
  */
 
 /**
@@ -27,12 +27,12 @@
  * Do not edit the class manually.
  */
 
-namespace Irisnet\APIV1\Client\Model;
+namespace Irisnet\APIV2\Client\Model;
 
 /**
  * Interface abstracting model access.
  *
- * @package Irisnet\APIV1\Client\Model
+ * @package Irisnet\APIV2\Client\Model
  * @author  OpenAPI Generator team
  */
 interface ModelInterface
@@ -93,4 +93,20 @@ interface ModelInterface
      * @return bool
      */
     public function valid();
+
+    /**
+     * Checks if a property is nullable
+     *
+     * @param string $property
+     * @return bool
+     */
+    public static function isNullable(string $property): bool;
+
+    /**
+     * Checks if a nullable property is set to null.
+     *
+     * @param string $property
+     * @return bool
+     */
+    public function isNullableSetToNull(string $property): bool;
 }
