@@ -3,17 +3,23 @@
  * @package IrisnetAPIClient
  */
 
+namespace Inc;
+
 use \GuzzleHttp\Client;
+use \Exception;
 use Inc\Helper\RulesHelper;
-use \GuzzleHttp\Cookie\CookieJar;
-use \Irisnet\APIV1\Client\ApiException;
-use \Irisnet\APIV1\Client\Model\IrisNet;
-use \Irisnet\APIV1\Client\Model\INParam;
-use \Irisnet\APIV1\Client\Model\INParams;
-use \Irisnet\APIV1\Client\Model\INDefault;
-use \Irisnet\APIV1\Client\Api\EndpointsToSetupTheAIApi;
-use \Irisnet\APIV1\Client\Api\EndpointsForAIChecksApi;
-use \Irisnet\APIV1\Client\Api\MiscellaneousOperationsApi;
+use \Irisnet\APIV2\Client\Model\Pricing;
+use \Irisnet\APIV2\Client\Model\CheckResult;
+use \Irisnet\APIV2\Client\ApiException;
+use \Irisnet\APIV2\Client\Configuration as APIConfiguration;
+use \Irisnet\APIV2\Client\Api\ConfigurationManagementApi;
+use \Irisnet\APIV2\Client\Api\DetailedConfigurationParametersApi;
+use \Irisnet\APIV2\Client\Api\BalanceEndpointsApi;
+use \Irisnet\APIV2\Client\Api\AICheckOperationsApi;
+use \Irisnet\APIV2\Client\Model\LicenseInfo;
+use \Irisnet\APIV2\Client\Model\Config;
+use \Irisnet\APIV2\Client\Model\Param;
+use \Irisnet\APIV2\Client\Model\ParamSet;
 
 /**
  * Helper class to connect to the irisnet API.
@@ -21,28 +27,173 @@ use \Irisnet\APIV1\Client\Api\MiscellaneousOperationsApi;
 class IrisnetAPIConnector
 {
 
-    private static $cookieJar;
-
     /**
-     * Sets the rule set for the upcoming checks.
+     * Sets the rules that should be used for the upcoming checks.
      *
-     * @param string $rule the given name of the rule set.
-     * @throws IrisnetException is thrown in case that the rule name could not be found or if the API request fails (will contain the status code and the message returned from the failed API request).
-     * @return boolean TRUE if the rule set is successfully applied.
+     * @param string|array $rule the given name of the rule set.
+     * @param string $license the license key to use. Omit if the next available license key should be used.
+     * @throws IrisnetException is thrown in case that there are no active licenses or if the API request fails (will contain the status code and the message returned from the failed API request).
+     * @return array the 'id' of the created configuration and the 'license' used to create the configuration.
      */
-    public static function setRules(string $rule) : bool
+    public static function setConfig(string|array $rule, $license = null) : array
     {
-        $parameters = self::createParameterModel($rule);
+        if ($license == null) {
+            $licenses = self::getActiveLicenses();
+            if ( empty($licenses) )
+                throw new Exception("No active license keys found. Try adding new licenses or activate existing once.");
+            $license = $licenses[0]['license'];
+        }
 
+        /** @var Config $config */
+        $config = IrisnetAPIConnector::createConfigModel($rule);
+        
+        /** @var ParamSet $params */
+        $params = IrisnetAPIConnector::createParameterModel($rule);
+
+        // Configure API key authorization: LICENSE-KEY
+        $apiConfig = APIConfiguration::getDefaultConfiguration()->setApiKey('LICENSE-KEY', $license);
+
+        $id = null;
         try {
-            $apiInstance = new EndpointsToSetupTheAIApi(
-                new Client(self::getClientConfig(true, true))
+            $apiInstance = new ConfigurationManagementApi(
+                null, // using default `GuzzleHttp\Client`
+                $apiConfig
             );
 
-            $apiInstance->setINParams($parameters);
+            /** @var Config $result */
+            $result = $apiInstance->setConfig($config);
+
+            $id = $result->getId();
+        } catch (ApiException $e) {
+            throw new IrisnetException("An Exception occurred while performing the API request '[POST] /v2/config'. ApiResponse: " . $e->getMessage(), $e->getCode());
+        }
+
+        if ($id === null)
+            throw new IrisnetException("The configuration could not be created. The API returned an empty id.");
+            
+        try {
+            $apiInstance = new DetailedConfigurationParametersApi(
+                null, // using default `GuzzleHttp\Client`
+                $apiConfig
+            );
+            $apiInstance->setParameters($id, $params);
+        } catch (ApiException $e) {
+            throw new IrisnetException("An Exception occurred while performing the API request '[POST] /v2/config/parameters'. ApiResponse: " . $e->getMessage(), $e->getCode());
+        }
+
+        return array(
+            'id' => $id,
+            'license' => $license
+        );
+    }
+
+    /**
+     * Deletes the AI configuration
+     *
+     * @param string $id the id of the configuration
+     * @param string $license the license key to use. Omit if the next available license key should be used.
+     * @throws IrisnetException is thrown in case that there are no active licenses or if the API request fails (will contain the status code and the message returned from the failed API request).
+     * @return bool the id of the created created configuration.
+     */
+    public static function deleteConfig(string $id, $license = null) : bool
+    {
+        if ($license == null) {
+            $licenses = self::getActiveLicenses();
+            if ( empty($licenses) )
+                throw new Exception("No active license keys found. Try adding new licenses or activate existing once.");
+            $license = $licenses[0]['license'];
+        }
+
+        // Configure API key authorization: LICENSE-KEY
+        $apiConfig = APIConfiguration::getDefaultConfiguration()->setApiKey('LICENSE-KEY', $license);
+
+        try {
+            $apiInstance = new ConfigurationManagementApi(
+                null, // using default `GuzzleHttp\Client`
+                $apiConfig
+            );
+
+            $result = $apiInstance->deleteConfig($id);
             return true;
         } catch (ApiException $e) {
-            throw new IrisnetException("An Exception occurred while performing the API request '/v1/setParameters'. ApiResponse: " . $e->getMessage(), $e->getCode());
+            throw new IrisnetException("An Exception occurred while performing the API request '[DELETE] /v2/config'. ApiResponse: " . $e->getMessage(), $e->getCode());
+        }
+    }
+
+    /**
+     * Retrieves the config and the parameters for the given config id and license key.
+     *
+     * @ignore
+     * @param string $id the id of the configuration
+     * @param string $license the license key where the id belongs to.
+     * @throws IrisnetException is thrown in case that the given rule name could not be found or if the API request fails (will contain the status code and the message returned from the failed API request).
+     * @return array the config and the parameters for the given config id and license key.
+     */
+    public static function getConfig(string $id, string $license) : array
+    {
+        // Configure API key authorization: LICENSE-KEY
+        $apiConfig = APIConfiguration::getDefaultConfiguration()->setApiKey('LICENSE-KEY', $license);
+
+        /** @var Config $config */
+        $config = null;
+
+        /** @var ParamSet $params */
+        $params = null;
+
+        try {
+            $apiInstance = new ConfigurationManagementApi(
+                null, // using default `GuzzleHttp\Client`
+                $apiConfig
+            );
+
+            $config = $apiInstance->getConfig($id);
+        } catch (ApiException $e) {
+            throw new IrisnetException("An Exception occurred while performing the API request '[GET] /v2/config/{configId}'. ApiResponse: " . $e->getMessage(), $e->getCode());
+        }
+
+        try {
+            $apiInstance = new DetailedConfigurationParametersApi(
+                null, // using default `GuzzleHttp\Client`
+                $apiConfig
+            );
+
+            $params = $apiInstance->getParameters($id);
+        } catch (ApiException $e) {
+            throw new IrisnetException("An Exception occurred while performing the API request '[GET] /v2/config/parameters/{configId}'. ApiResponse: " . $e->getMessage(), $e->getCode());
+        }
+
+        return array(
+            'prototypes' => $config,
+            'paramSet' => $params
+        );
+    }
+
+    /**
+     * Determines the cost in credits of the rule set.
+     *
+     * @ignore
+     * @param string $id the id of the configuration
+     * @param string $license the license key where the id belongs to.
+     * @throws IrisnetException is thrown in case that the given rule name could not be found or if the API request fails (will contain the status code and the message returned from the failed API request).
+     * @return integer
+     */
+    public static function getCost(string $id, string $license) : int
+    {
+        // Configure API key authorization: LICENSE-KEY
+        $apiConfig = APIConfiguration::getDefaultConfiguration()->setApiKey('LICENSE-KEY', $license);
+
+        try {
+            $apiInstance = new BalanceEndpointsApi(
+                null, // using default `GuzzleHttp\Client`
+                $apiConfig
+            );
+
+            /** @var Pricing $pricing */
+            $pricing = $apiInstance->getCost($id);
+
+            return $pricing->getCost();
+        } catch (ApiException $e) {
+            throw new IrisnetException("An Exception occurred while performing the API request '[GET] /v2/cost/{configId}'. ApiResponse: " . $e->getMessage(), $e->getCode());
         }
     }
     
@@ -50,21 +201,27 @@ class IrisnetAPIConnector
      * Makes the API call to check the image for the specified rules.
      *
      * @param string $file the name (including path) of the image or url of an image that needs to be checked. 
+     * @param string $rule the given name of the rule set.
      * @param integer $detail Sets the response details. Use 1 (default) for minimum detail (better API performance), 2 for medium details and 3 for all details.
-     * @param string $rule the given name of the rule set. Omit if the cost of the last set rule set should be determined.
      * @param integer $licenseId the id of the license key to use. Omit if the next available license key should be used.
      * @throws IrisnetException is thrown in case that the rule name could not be found, if the license id does not exist, 
      * if the license key is not active or out of credits, if the specified filename does not exist 
      * or if the API request fails (will contain the status code and the message returned from the failed API request).
-     * @return IrisNet Contains information on the AI result from the source media check. See <a href="https://www.irisnet.de/api/">API page</a> for more information
+     * @return CheckResult Contains information on the AI result from the source media check. See <a href="https://www.irisnet.de/api/">API page</a> for more information
      */
-    public static function processImage(string $file, int $detail = 1, string $rule = null, int $licenseId = null) : IrisNet
+    public static function checkImage(string $file, string $rule, int $detail = 1, int $licenseId = null) : CheckResult
     {
 
         $isUrl = filter_var($file, FILTER_VALIDATE_URL);
 
         if (!$isUrl && !file_exists($file)) {
             throw new IrisnetException("The specified file does not exist. Filename: $file", 404);
+        }
+
+        if (!$isUrl) {
+            $type = pathinfo($file, PATHINFO_EXTENSION);
+            $data = file_get_contents($file);
+            $base64 = 'data:image/' . $type . ';base64,' . base64_encode($data);
         }
         
         $usable = self::getUsableLicenses();
@@ -80,71 +237,21 @@ class IrisnetAPIConnector
         }
         $key = $license['license'];
 
-        if ($rule !== null) {
-            self::setRules($rule);
-        }
+        $rule = self::getRuleOption($rule);
+        $configId = $rule['id'];
 
-        $apiInstance = new EndpointsForAIChecksApi(
-            new Client(self::getClientConfig(true))
+        // Configure API key authorization: LICENSE-KEY
+        $apiConfig = APIConfiguration::getDefaultConfiguration()->setApiKey('LICENSE-KEY', $key);
+
+        $apiInstance = new AICheckOperationsApi(
+            null, // using default `GuzzleHttp\Client`
+            $apiConfig
         );
 
         try {
-            if (!$isUrl)
-                return $apiInstance->checkImage($key, new \SplFileObject($file), $detail, true);
-
-            return $apiInstance->checkImageUrl($file, $key, $detail, true);
+            return $apiInstance->checkImage($configId, null, isset($base64) ? $base64 : $file, $detail, true);
         } catch (ApiException $e) {
-            throw new IrisnetException("An Exception occurred while performing the API request '/v1/check-image' or '/v1/check-url'. ApiResponse: " . $e->getMessage(), $e->getCode());
-        }
-    }
-
-    /**
-     * Downloads the modified image as specified by the rules parameters, if needed.
-     *
-     * @param string $filename the name of the file (without path) that should be downloaded. Is equal to the file name that was processed.
-     * @param string $downloadPath the location (folder) of where to save the downloaded file. When omitted the return type is SplFileObject
-     * @throws IrisnetException is thrown in case that the API request fails (will contain the status code and the message returned from the failed API request).
-     * @return boolean|\SplFileObject returns FALSE in case that the file could not be saved at the specified location
-     */
-    public static function getProcessedImage(string $filename, string $downloadPath = null) {
-        try {
-            $apiInstance = new MiscellaneousOperationsApi(
-                new Client(self::getClientConfig(true))
-            );
-
-            $file = $apiInstance->downloadProcessed($filename);
-
-            if ($downloadPath == null)
-                return $file;
-
-            return !file_put_contents($downloadPath . $filename, $file->fread($file->getSize())) ? false : true;
-        } catch (ApiException $e) {
-            throw new IrisnetException("An Exception occurred while performing the API request '/v1/download'. ApiResponse: " . $e->getMessage(), $e->getCode());
-        }
-    }
-
-    /**
-     * Determines the cost in credits of the rule set.
-     *
-     * @ignore
-     * @param string $rule the given name of the rule set. Omit if the cost of the last set rule set should be determined.
-     * @throws IrisnetException is thrown in case that the given rule name could not be found or if the API request fails (will contain the status code and the message returned from the failed API request).
-     * @return integer
-     */
-    public static function getCost(string $rule = null) : int
-    {
-        if ($rule !== null) {
-            $parameters = self::setRules($rule);
-        }
-
-        try {
-            $apiInstance = new MiscellaneousOperationsApi(
-                new Client(self::getClientConfig(true))
-            );
-
-            return intval($apiInstance->getAICost());
-        } catch (ApiException $e) {
-            throw new IrisnetException("An Exception occurred while performing the API request '/v1/cost'. ApiResponse: " . $e->getMessage(), $e->getCode());
+            throw new IrisnetException("An Exception occurred while performing the API request '/v2/check-image/{configId}'. ApiResponse: " . $e->getMessage(), $e->getCode());
         }
     }
 
@@ -157,12 +264,8 @@ class IrisnetAPIConnector
      */
     public static function refreshCredits()
     {
-        $options = get_option('irisnet_plugin_licenses');
+        $options = get_option('irisnet_plugin_licenses');       
 
-        $apiInstance = new MiscellaneousOperationsApi(
-            new Client(self::getClientConfig())
-        );
-        
         $refreshCount = 0;
         $deactivatedCount = 0;
         foreach ($options as $key => $value) {
@@ -171,9 +274,17 @@ class IrisnetAPIConnector
             }
 
             $license = $value['license'];
+            
+            // Configure API key authorization: LICENSE-KEY
+            $apiConfig = APIConfiguration::getDefaultConfiguration()->setApiKey('LICENSE-KEY', $license);
+            $apiInstance = new BalanceEndpointsApi(
+                null, // using default `GuzzleHttp\Client`
+                $apiConfig
+            );
 
             try {
-                $result = $apiInstance->getLicenseInfo($license);
+                /** @var LicenseInfo $result */
+                $result = $apiInstance->getLicenseInfo();
 
                 $options[$key]['credits_used'] = $result->getCreditsUsed();
                 $options[$key]['total_credits'] = $result->getTotalCredits();
@@ -221,9 +332,9 @@ class IrisnetAPIConnector
         $options = get_option('irisnet_plugin_licenses');
 
         $active = array();
-        foreach ($options as $key => $value) {
+        foreach ($options as $value) {
             if (isset($value['is_active']) && $value['is_active'] == true) {
-                $active[$key] = $value;
+                $active[] = $value;
             }
         }
 
@@ -253,15 +364,124 @@ class IrisnetAPIConnector
     }
 
     /**
+     * Creates a prototypes config model from the given rule set or rule set name.
+     *
+     * @ignore
+     * @param string|array $rule the given name of the rule set or rule set array.
+     * @throws IrisnetException is thrown in case that the given rule name could not be found.
+     * @throws InvalidArgumentException if the given parameter is not a valid type (string or array)
+     * @return Config Config model with prototypes as it is consumed by the Irisnet API
+     */
+    public static function createConfigModel($rule) : Config
+    {
+        $rule = self::getRuleOption($rule);
+
+        // get all prototypes
+        $prototypes = array_keys(RulesHelper::getSimplifiedClassObjectArray());
+
+        // filter parameters used
+        $parameters = array_unique(array_map(function ($v) {
+            return substr($v, 0, strpos($v, '_'));
+        }, array_keys($rule)));
+
+        // change prototype 'baseParameters' to 'nudityCheck'
+        // keep for downward compatibility
+        $prototypes = array_map(function($v) {
+            return $v === 'baseParameters' ? 'nudityCheck' : $v;
+        }, $prototypes);
+        $parameters = array_map(function($v) {
+            return $v === 'baseParameters' ? 'nudityCheck' : $v;
+        }, $parameters);
+
+        // filter $prototypes for elements that are in both arrays
+        $prototypes = array_intersect($prototypes, $parameters);
+
+        // return prototypes config
+        return new Config(array('prototypes' => array_values($prototypes)));
+    }
+
+    /**
      * Creates a parameter model from the given rule set or rule set name.
      *
      * @ignore
      * @param string|array $rule the given name of the rule set or rule set array.
      * @throws IrisnetException is thrown in case that the given rule name could not be found.
-     * @throws IrisnetException if the given parameter is not a valid type (string or array)
-     * @return INParams Parameter model as it is consumed by the Irisnet API
+     * @throws InvalidArgumentException if the given parameter is not a valid type (string or array)
+     * @return ParamSet Parameter set model as it is consumed by the Irisnet API
      */
-    public static function createParameterModel($rule) : INParams
+    public static function createParameterModel($rule) : ParamSet
+    {
+        $rule = self::getRuleOption($rule);
+
+        // get all prototypes
+        $prototypes = array_keys(RulesHelper::getSimplifiedClassObjectArray());
+
+        // remove classification names that are also prototype names
+        $prototypes = array_filter($prototypes, function($v) {
+            return RulesHelper::findClassParent($v) === false;
+        });
+
+        // filter parameters used
+        $parameters = array_unique(array_map(function ($v) {
+            return substr($v, 0, strpos($v, '_'));
+        }, array_keys($rule)));
+
+        // remove filtered prototype values from parameters
+        $parameters = array_diff($parameters, $prototypes, array('baseParameters', 'default') /* keep 'baseParameters' for downward compatibility */);
+
+        // remove elements ending with _switch from $options
+        $rule = array_filter($rule, function($v, $k) {
+            return strpos($k, '_switch') === false;
+        }, ARRAY_FILTER_USE_BOTH);
+
+        // loop through parameters and create Prarameter objects
+        $params = array();
+        foreach ($parameters as $parameter) {
+            // filter $rule array for elements with keys starting with $parameter
+            $paramData = array_filter($rule, function($k) use ($parameter) {
+                return strpos($k, $parameter) === 0;
+            }, ARRAY_FILTER_USE_KEY);
+
+            // remove first part of the key matching the $parameter string
+            $tmp = array();
+            array_walk($paramData, function($v, $k) use (&$tmp, $parameter) {
+                $tmp[substr($k, strlen($parameter) + 1)] = $v;
+            });
+            $paramData = $tmp;
+            
+            // add min/max (0, 0) values if missing and allowMinMax is false
+            $allowMinMax = RulesHelper::getClassObjectGroups(true)[RulesHelper::findClassParent($parameter)][$parameter]['allowMinMax'];
+            if ( ! $allowMinMax ) {
+                if ( ! array_key_exists('min', $paramData) ) {
+                    $paramData['min'] = 0;
+                }
+                if ( ! array_key_exists('max', $paramData) ) {
+                    $paramData['max'] = 0;
+                }
+            }
+
+            $paramData['classification'] = $parameter;
+
+            // create Param object
+            $params[] = new Param($paramData);
+        }
+
+
+        // create ParamSet object
+        /** @var ParamSet $paramSet */
+        $paramSet = new ParamSet(array('params' => $params));
+
+        // set the default values
+        if ( isset($rule['default_thresh']) )
+            $paramSet->setThresh($rule['default_thresh']);
+        if ( isset($rule['default_grey']) )
+            $paramSet->setGrey($rule['default_grey']);
+
+        
+        return $paramSet;
+    }
+
+    public static function getRuleOption($rule) : array
     {
         if(is_array($rule) && isset($rule['rule_name'])) {
             $option = $rule;
@@ -273,7 +493,7 @@ class IrisnetAPIConnector
     
             $option = $options[$rule];
         } else {
-            throw new InvalidArgumentException("Argument not of expected type. Please provide the name (string) of a rule set.");
+            throw new \InvalidArgumentException("Argument not of expected type. Please provide the name (string) of a rule set.");
         }
 
         // unset some variables that are not needed in this scope
@@ -281,116 +501,6 @@ class IrisnetAPIConnector
         unset($option['description']);
         unset($option['cost']);
 
-        // create INParams object
-        $params = new INParams();
-
-        // create and fill INDefault object
-        $keys = array_keys($option);
-        $default = new INDefault();
-        foreach ($keys as $key) {
-            $exploded = explode('_', $key, 2);
-            if ($exploded[0] === 'default' && $exploded[1] !== 'switch') {
-                $default->{'set' . ucfirst($exploded[1])}($option[$key]);
-                unset($option[$key]);
-            }
-        }
-        $params->setInDefault($default);
-
-        // create a class object array
-        $classObjects = array();
-        foreach (array_values(RulesHelper::getClassObjectGroups()) as $group) {
-            $classObjects = array_merge($classObjects, array_keys($group));
-        }
-
-        // create a group object array
-        $groupObjects = array();
-        foreach (RulesHelper::getClassObjectGroups(true) as $key => $value) {
-            $group = array();
-            foreach ($value as $className => $classOptions) {
-                if (!$classOptions['allowMinMax'])
-                    $group[] = $className;
-            }
-            $groupObjects[$key] = $group;
-        }
-
-        // create and fill INParam objects as needed
-        $paramArray = array();
-        foreach ($classObjects as $class) {
-            $data = array();
-            foreach ($option as $key => $value) {
-                $exploded = explode('_', $key, 2);
-                if (in_array($exploded[0], array_keys($groupObjects)) && $exploded[1] === 'switch') {
-                    continue;
-                }
-
-                if ($exploded[0] === $class) {
-                    if (in_array($exploded[0], array_values($classObjects)) && ($exploded[1] === 'param_switch')) {
-                        $data['min'] = '0';
-                        $data['max'] = '0';
-                    } else {
-                        $data[$exploded[1]] = $value;
-                    }
-                    unset($option[$key]);
-                }
-            }
-
-            if (!empty($data)) {
-                $data['in_class'] = $class;
-                $paramArray[] = new INParam($data);
-            }
-        }
-
-        // Check paramArray for missing inParams from remaining groups within $option. 
-        // get all created param classes
-        $names = array();
-        foreach($paramArray as $param) {
-            $names[] = $param->getInClass();
-        };
-
-        // flatten groupObjects and find missing from paramArray
-        $res = array();
-        array_walk_recursive($groupObjects, function($v) use (&$res) { $res[] = $v; });
-        $missing = array_diff($res, $names);
-
-        // Create a new inParam for each class with param_switch that was not active: Use data max = -1; draw_mode = 0
-        foreach ($option as $key => $value) {
-            $groupName = explode('_', $key, 2)[0];
-
-            foreach ($missing as $c) {
-                if (!in_array($c, $groupObjects[$groupName]))
-                    continue;
-
-                $paramArray[] = new INParam(array(
-                    'in_class' => $c,
-                    'max' => -1,
-                    'draw_mode' => 0
-                ));
-            }
-        }
-        
-        // fill INParams object
-        $params->setInParam($paramArray);
-
-        return $params;
-    }
-
-    private static function getClientConfig($withCookie = false, $clearOldCookie = false) : array {
-        if ($clearOldCookie && self::$cookieJar) {
-            self::$cookieJar->clear();
-        }
-
-        if (!self::$cookieJar && $withCookie) {
-            self::$cookieJar = new CookieJar();
-        }
-
-        $arr = array();
-
-        if ($withCookie && self::$cookieJar) {
-            $arr['cookies'] = self::$cookieJar;
-        }
-
-        // add further config args here
-
-        return $arr;
+        return $option;
     }
 }
